+++ panel/src/mixins/section.js
